+++ vega-dataflow/src/main/kotlin/dev/aster/vega/dataflow/transform/Transform.kt
@@ -304,6 +304,7 @@ public class TransformRegistry(transforms: List<Transform>) {
           ResolveFilterTransform,
           IsocontourTransform,
           ContourTransform,
+          WordcloudTransform,
           GeoPathTransform,
           Kde2dTransform,
           HeatmapTransform,